refactor: move hooks under `hooks` directory
