+++ src/components/Flowbite/Flowbite.spec.tsx
@@ -1,7 +1,7 @@
 import { act, render } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
 import { mergeDeep } from '../../helpers/merge-deep';
-import { useThemeMode } from '../../helpers/use-theme-mode';
+import { useThemeMode } from '../../hooks/use-theme-mode';
 import { getTheme } from '../../theme-store';
 import { Flowbite } from '../Flowbite';
